+++ src/Storefront/Data/Struct/AbstractCheckoutData.php
@@ -8,21 +8,10 @@
 namespace Swag\PayPal\Storefront\Data\Struct;
 
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Core\Framework\Struct\Struct;
 
 #[Package('checkout')]
-class AbstractCheckoutData extends Struct
+class AbstractCheckoutData extends AbstractScriptData
 {
-    protected string $clientId;
-
-    protected string $merchantPayerId;
-
-    protected string $languageIso;
-
-    protected string $currency;
-
-    protected string $intent;
-
     protected string $buttonShape;
 
     protected string $buttonColor;
@@ -49,56 +38,6 @@ class AbstractCheckoutData extends Struct
     protected ?string $accountOrderEditFailedUrl = null;
 
     protected string $brandName;
-
-    public function getClientId(): string
-    {
-        return $this->clientId;
-    }
-
-    public function setClientId(string $clientId): void
-    {
-        $this->clientId = $clientId;
-    }
-
-    public function getMerchantPayerId(): string
-    {
-        return $this->merchantPayerId;
-    }
-
-    public function setMerchantPayerId(string $merchantPayerId): void
-    {
-        $this->merchantPayerId = $merchantPayerId;
-    }
-
-    public function getLanguageIso(): string
-    {
-        return $this->languageIso;
-    }
-
-    public function setLanguageIso(string $languageIso): void
-    {
-        $this->languageIso = $languageIso;
-    }
-
-    public function getCurrency(): string
-    {
-        return $this->currency;
-    }
-
-    public function setCurrency(string $currency): void
-    {
-        $this->currency = $currency;
-    }
-
-    public function getIntent(): string
-    {
-        return $this->intent;
-    }
-
-    public function setIntent(string $intent): void
-    {
-        $this->intent = $intent;
-    }
 
     public function getUserIdToken(): ?string
     {